Create ViewAddDigits
Create ViewAddDigits to store all the last lesson challenge into this View
